code/requirements.txt
streamlit==1.30.0
openai==1.7.2
matplotlib==3.8.2
plotly==5.18.0
scipy==1.11.4
scikit-learn==1.3.2
transformers==4.36.2
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-ai-contentsafety==1.0.0b1
requests==2.31.0
tiktoken==0.5.2
azure-storage-queue==12.9.0
langchain==0.1.0
langchain-community==0.0.12
langchain-openai==0.0.2.post1
beautifulsoup4==4.12.2
fake-useragent==1.4.0
chardet==5.2.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0b8
opencensus-ext-azure==1.1.13
pandas==1.5.1
python-docx==0.8.11